% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Bar chart}
\usage{
barchart(
  data,
  title,
  magnitude,
  xlab,
  ylab,
  source_description,
  source_url,
  palette,
  patterns,
  width = "100\%",
  height = "auto",
  elementId = NULL
)
}
\arguments{
\item{data}{Data frame containing data to plot}

\item{title}{chart title}

\item{magnitude}{0 for counts, 3 for thousands, etc.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{source_description}{text for data source}

\item{source_url}{url for data source}

\item{palette}{bar colours}

\item{patterns}{bar patterns}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Block ID}
}
\description{
Bar chart
}
\examples{
""
}
