% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcdata}
\alias{bcdata}
\title{bcdata}
\format{
\subsection{\code{bcdata}}{

A data frame with 124 rows and 60 columns:
\describe{
\item{g1}{grouping variable}
\item{g2}{outcome}
\item{value}{value (bar height / count)}
\item{lo}{lower bound}
\item{hi}{upper bound}
}
}
}
\usage{
bcdata
}
\description{
A  dataset appropriate for use with \code{barchart}
}
\keyword{datasets}
