% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circletab.R
\name{circletab}
\alias{circletab}
\title{Table with circles}
\usage{
circletab(
  data,
  max_value,
  fill_color = "rgba(59, 130, 246, .5)",
  stroke_width = 2,
  stroke_color = "#333",
  width = "100\%",
  height = "auto",
  elementId = NULL
)
}
\arguments{
\item{data}{Data frame with three columns with two categorical variables and a single measure (count, proportion, etc.)}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Block ID}

\item{max_val}{Lower and upper values for bar track}

\item{bar_color}{Bar color}

\item{track_color}{Track color}

\item{font_color}{Font color}
}
\description{
Table where values are represented as circles of varying size.
}
\examples{
set.seed(1234)

d <- data.frame(
  g1 = rep(c("Foo", "Bar", "Baz"), each = 3),
  g2 = rep(c("A", "B", "C"), 3),
  value = runif(9)
)

circletab(d, max_value = 1)
}
