% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linechart.R
\name{linechart}
\alias{linechart}
\title{Line chart}
\usage{
linechart(
  data,
  title,
  magnitude,
  xlab,
  ylab,
  source_description,
  source_url,
  palette,
  linetype,
  width = "100\%",
  height = "auto",
  elementId = NULL
)
}
\arguments{
\item{data}{Data frame containing data to plot}

\item{title}{chart title}

\item{magnitude}{0 for counts, 3 for thousands, etc.}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{source_description}{text for data source}

\item{source_url}{url for data source}

\item{palette}{bar colours}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{Block ID}

\item{patterns}{bar patterns}
}
\description{
Line chart
}
\examples{
linechart(tsdata)
}
