% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{reshape_bcdata}
\alias{reshape_bcdata}
\title{Format dataset}
\usage{
reshape_bcdata(data)
}
\arguments{
\item{data}{data Data frame with three columns with two categorical variables and a single measure (count, proportion, etc.)}
}
\description{
Format dataset
}
