% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circletab.R
\name{reshape_ctdata}
\alias{reshape_ctdata}
\title{Format dataset}
\usage{
reshape_ctdata(data)
}
\arguments{
\item{data}{data Data frame with three columns with two categorical variables and a single measure (count, proportion, etc.)}
}
\description{
Format dataset
}
