% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linechart.R
\name{reshape_lcdata}
\alias{reshape_lcdata}
\title{Format dataset}
\usage{
reshape_lcdata(data)
}
\arguments{
\item{data}{data Data frame with three columns with two categorical variables and a single measure (count, proportion, etc.)}
}
\description{
Format dataset
}
