% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsdata}
\alias{tsdata}
\title{tsdata}
\format{
\subsection{\code{tsdata}}{

A data frame with 124 rows and 60 columns:
\describe{
\item{x}{date (x-axis data)}
\item{variabe}{grouping variable for line colors}
\item{y}{value (y-axis data)}
\item{lo}{lower bound}
\item{hi}{upper bound}
}
}
}
\usage{
tsdata
}
\description{
A randomly created dataset appropriate for use with \code{linechart}
}
\keyword{datasets}
